поменял названия класса и интерфейса
